pyproject.toml
[build-system]
requires = ["maturin>=1.7,<2.0"]
build-backend = "maturin"

[project]
name = "three_dev"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]
dependencies = [
    "customtkinter>=5.2.2",
    "pillow>=10.4.0",
]
[project.optional-dependencies]
tests = [
    "pytest",
]
[tool.maturin]
python-source = "python"
features = ["pyo3/extension-module"]
